approvals: Add check-approvals and approvaltest
Adds two new tools, `check-approvals` and an `approvaltest` package. The
latter allows clients to store a receive events and store them in json
`.received.json` suffixed files, while the former checks if the received
files match the `.approved.json` suffixed files, and prompts whether the
changes want to be accepted or not.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>